requirements.txt
python-telegram-bot==13.7
environs==14.3.0
google-cloud-dialogflow==2.41.2
vk-api==11.10.0